Add css for borders, move colours location
